🐛 Fix: 로그인 안 됐을 때 logo 안 뜨는 문제 해결
